Move string-related utility to functions to strings.py, const.py
