+++ src/black/__init__.py
@@ -42,12 +42,18 @@ from typing import (
     cast,
     TYPE_CHECKING,
 )
-from mypy_extensions import mypyc_attr
 
-from appdirs import user_cache_dir
 from dataclasses import dataclass, field, replace
 import click
 import toml
+
+from black.const import DEFAULT_LINE_LENGTH, DEFAULT_INCLUDES, DEFAULT_EXCLUDES
+from black.const import CACHE_DIR, STDIN_PLACEHOLDER, STRING_PREFIX_CHARS
+from black.strings import sub_twice, re_compile_maybe_verbose
+from black.strings import format_hex, format_scientific_notation
+from black.strings import format_long_or_complex_number, format_float_or_int_string
+from black.strings import dump_to_file, diff, color_diff
+from black.strings import has_triple_quotes, fix_docstring
 
 try:
     from typed_ast import ast3, ast27
@@ -80,14 +86,6 @@ else:
 
 if TYPE_CHECKING:
     import colorama  # noqa: F401
-
-DEFAULT_LINE_LENGTH = 88
-DEFAULT_EXCLUDES = r"/(\.direnv|\.eggs|\.git|\.hg|\.mypy_cache|\.nox|\.tox|\.venv|venv|\.svn|_build|buck-out|build|dist)/"  # noqa: B950
-DEFAULT_INCLUDES = r"\.pyi?$"
-CACHE_DIR = Path(user_cache_dir("black", version=__version__))
-STDIN_PLACEHOLDER = "__BLACK_STDIN_FILENAME__"
-
-STRING_PREFIX_CHARS: Final = "furbFURB"  # All possible string prefix characters.
 
 
 # types
@@ -943,22 +941,6 @@ def format_file_in_place(
             f.detach()
 
     return True
-
-
-def color_diff(contents: str) -> str:
-    """Inject the ANSI color codes to the diff."""
-    lines = contents.split("\n")
-    for i, line in enumerate(lines):
-        if line.startswith("+++") or line.startswith("---"):
-            line = "\033[1;37m" + line + "\033[0m"  # bold white, reset
-        elif line.startswith("@@"):
-            line = "\033[36m" + line + "\033[0m"  # cyan, reset
-        elif line.startswith("+"):
-            line = "\033[32m" + line + "\033[0m"  # green, reset
-        elif line.startswith("-"):
-            line = "\033[31m" + line + "\033[0m"  # red, reset
-        lines[i] = line
-    return "\n".join(lines)
 
 
 def wrap_stream_for_windows(
@@ -4746,15 +4728,6 @@ def insert_str_child_factory(string_leaf: Leaf) -> Callable[[LN], None]:
     return insert_str_child
 
 
-def has_triple_quotes(string: str) -> bool:
-    """
-    Returns:
-        True iff @string starts with three quotation characters.
-    """
-    raw_string = string.lstrip(STRING_PREFIX_CHARS)
-    return raw_string[:3] in {'"""', "'''"}
-
-
 def parent_type(node: Optional[LN]) -> Optional[NodeType]:
     """
     Returns:
@@ -5381,46 +5354,6 @@ def normalize_numeric_literal(leaf: Leaf) -> None:
     else:
         text = format_float_or_int_string(text)
     leaf.value = text
-
-
-def format_hex(text: str) -> str:
-    """
-    Formats a hexadecimal string like "0x12B3"
-    """
-    before, after = text[:2], text[2:]
-    return f"{before}{after.upper()}"
-
-
-def format_scientific_notation(text: str) -> str:
-    """Formats a numeric string utilizing scentific notation"""
-    before, after = text.split("e")
-    sign = ""
-    if after.startswith("-"):
-        after = after[1:]
-        sign = "-"
-    elif after.startswith("+"):
-        after = after[1:]
-    before = format_float_or_int_string(before)
-    return f"{before}e{sign}{after}"
-
-
-def format_long_or_complex_number(text: str) -> str:
-    """Formats a long or complex string like `10L` or `10j`"""
-    number = text[:-1]
-    suffix = text[-1]
-    # Capitalize in "2L" because "l" looks too similar to "1".
-    if suffix == "l":
-        suffix = "L"
-    return f"{format_float_or_int_string(number)}{suffix}"
-
-
-def format_float_or_int_string(text: str) -> str:
-    """Formats a float string like "1.0"."""
-    if "." not in text:
-        return text
-
-    before, after = text.split(".")
-    return f"{before or 0}.{after or 0}"
 
 
 def normalize_invisible_parens(node: Node, parens_after: Set[str]) -> None:
@@ -6580,19 +6513,6 @@ def assert_stable(src: str, dst: str, mode: Mode) -> None:
         ) from None
 
 
-@mypyc_attr(patchable=True)
-def dump_to_file(*output: str, ensure_final_newline: bool = True) -> str:
-    """Dump `output` to a temporary file. Return path to the file."""
-    with tempfile.NamedTemporaryFile(
-        mode="w", prefix="blk_", suffix=".log", delete=False, encoding="utf8"
-    ) as f:
-        for lines in output:
-            f.write(lines)
-            if ensure_final_newline and lines and lines[-1] != "\n":
-                f.write("\n")
-    return f.name
-
-
 @contextmanager
 def nullcontext() -> Iterator[None]:
     """Return an empty context manager.
@@ -6600,26 +6520,6 @@ def nullcontext() -> Iterator[None]:
     To be used like `nullcontext` in Python 3.7.
     """
     yield
-
-
-def diff(a: str, b: str, a_name: str, b_name: str) -> str:
-    """Return a unified diff string between strings `a` and `b`."""
-    import difflib
-
-    a_lines = [line for line in a.splitlines(keepends=True)]
-    b_lines = [line for line in b.splitlines(keepends=True)]
-    diff_lines = []
-    for line in difflib.unified_diff(
-        a_lines, b_lines, fromfile=a_name, tofile=b_name, n=5
-    ):
-        # Work around https://bugs.python.org/issue2142
-        # See https://www.gnu.org/software/diffutils/manual/html_node/Incomplete-Lines.html
-        if line[-1] == "\n":
-            diff_lines.append(line)
-        else:
-            diff_lines.append(line + "\n")
-            diff_lines.append("\\ No newline at end of file\n")
-    return "".join(diff_lines)
 
 
 def cancel(tasks: Iterable["asyncio.Task[Any]"]) -> None:
@@ -6653,26 +6553,6 @@ def shutdown(loop: asyncio.AbstractEventLoop) -> None:
         cf_logger = logging.getLogger("concurrent.futures")
         cf_logger.setLevel(logging.CRITICAL)
         loop.close()
-
-
-def sub_twice(regex: Pattern[str], replacement: str, original: str) -> str:
-    """Replace `regex` with `replacement` twice on `original`.
-
-    This is used by string normalization to perform replaces on
-    overlapping matches.
-    """
-    return regex.sub(replacement, regex.sub(replacement, original))
-
-
-def re_compile_maybe_verbose(regex: str) -> Pattern[str]:
-    """Compile a regular expression string in `regex`.
-
-    If it contains newlines, use verbose mode.
-    """
-    if "\n" in regex:
-        regex = "(?x)" + regex
-    compiled: Pattern[str] = re.compile(regex)
-    return compiled
 
 
 def enumerate_reversed(sequence: Sequence[T]) -> Iterator[Tuple[Index, T]]:
@@ -7038,52 +6918,6 @@ def is_docstring(leaf: Leaf) -> bool:
         return True
 
     return False
-
-
-def lines_with_leading_tabs_expanded(s: str) -> List[str]:
-    """
-    Splits string into lines and expands only leading tabs (following the normal
-    Python rules)
-    """
-    lines = []
-    for line in s.splitlines():
-        # Find the index of the first non-whitespace character after a string of
-        # whitespace that includes at least one tab
-        match = re.match(r"\s*\t+\s*(\S)", line)
-        if match:
-            first_non_whitespace_idx = match.start(1)
-
-            lines.append(
-                line[:first_non_whitespace_idx].expandtabs()
-                + line[first_non_whitespace_idx:]
-            )
-        else:
-            lines.append(line)
-    return lines
-
-
-def fix_docstring(docstring: str, prefix: str) -> str:
-    # https://www.python.org/dev/peps/pep-0257/#handling-docstring-indentation
-    if not docstring:
-        return ""
-    lines = lines_with_leading_tabs_expanded(docstring)
-    # Determine minimum indentation (first line doesn't count):
-    indent = sys.maxsize
-    for line in lines[1:]:
-        stripped = line.lstrip()
-        if stripped:
-            indent = min(indent, len(line) - len(stripped))
-    # Remove indentation (first line is special):
-    trimmed = [lines[0].strip()]
-    if indent < sys.maxsize:
-        last_line_idx = len(lines) - 2
-        for i, line in enumerate(lines[1:]):
-            stripped_line = line[indent:].rstrip()
-            if stripped_line or i == last_line_idx:
-                trimmed.append(prefix + stripped_line)
-            else:
-                trimmed.append("")
-    return "\n".join(trimmed)
 
 
 if __name__ == "__main__":
